Move provenance related code to separate file
